pyproject.toml
[project]
name = "news-fetcher"
version = "0.1.0"
description = "MCP server for fetching, processing, and serving news content with EPUB generation and OPDS support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "beautifulsoup4>=4.13.5",
    "ebooklib>=0.19",
    "fastapi>=0.117.1",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "lxml>=6.0.2",
    "mcp[cli]>=1.15.0",
    "pydantic>=2.11.9",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.5",
    "trafilatura>=1.6.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
